dashboard_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
jinja2
requests
python-multipart